example/transformers/ext/stable_diffusion/requirements.txt
torch >= 1.9
numpy
Pillow
regex
tqdm